// frontend/src/api/generated/models/raceDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */
import type { GoalPriority } from './goalPriority';
import type { SportType } from './sportType';

export interface RaceDto {
  id?: string;
  athleteUserId?: string;
  /** @nullable */
  seasonId?: string | null;
  /** @nullable */
  goalId?: string | null;
  /** @nullable */
  name?: string | null;
  sport?: SportType;
  startsAtUtc?: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  distanceMeters?: number | null;
  /** @nullable */
  elevationGainMeters?: number | null;
  priority?: GoalPriority;
  /** @nullable */
  targetTimeSeconds?: number | null;
  /** @nullable */
  targetResultNote?: string | null;
  /** @nullable */
  actualTimeSeconds?: number | null;
  /** @nullable */
  actualResultNote?: string | null;
  /** @nullable */
  resultNotes?: string | null;
}